requirements.txt
weaviate-client
llama-index
llama-index-vector-stores-weaviate
llama-index-embeddings-huggingface